bot/requirements.txt
certifi==2017.7.27.1
emoji==0.5.1
ephem==3.7.6.0
future==0.17.1
PySocks==1.6.8
python-telegram-bot==11.1.0